Subject : doc
Tag : javadoc

Impl : un peu de doc et nettoyage d'imports/ commentaires

notable issue :
