+++ src/main/java/fr/diginamic/qualiair/mapper/MesureAirMapper.java
@@ -2,11 +2,8 @@ package fr.diginamic.qualiair.mapper;
 
 import fr.diginamic.qualiair.dto.atmofrance.AirDataFeatureDto;
 import fr.diginamic.qualiair.dto.atmofrance.AirDataPropertiesDto;
-import fr.diginamic.qualiair.dto.insertion.CommuneHabitantDto;
 import fr.diginamic.qualiair.entity.MesureAir;
-import fr.diginamic.qualiair.entity.MesurePopulation;
 import fr.diginamic.qualiair.entity.TypeMesure;
-import fr.diginamic.qualiair.exception.ParsedDataException;
 import org.springframework.stereotype.Component;
 
 import java.time.LocalDate;
@@ -15,29 +12,12 @@ import java.util.ArrayList;
 import java.util.List;
 
 import static fr.diginamic.qualiair.utils.MesureUtils.cleanUpElementCode;
-import static fr.diginamic.qualiair.utils.MesureUtils.toInt;
 
 /**
- * Mesure Mapper
+ * Mapper pour les mesures qualité de l'air
  */
 @Component
-public class MesureMapper {
-
-    /**
-     * Map a dto from csv to entity
-     *
-     * @param dto dto from csv
-     * @return entity
-     */
-    public MesurePopulation toEntity(CommuneHabitantDto dto, LocalDateTime date) throws ParsedDataException {
-        MesurePopulation mesure = new MesurePopulation();
-        mesure.setTypeMesure(TypeMesure.RELEVE_POPULATION);
-        mesure.setDateReleve(date);
-        mesure.setDateEnregistrement(LocalDateTime.now());
-        mesure.setValeur(toInt(dto.getPopulationMunicipale().trim().replace(" ", "")));
-        return mesure;
-    }
-
+public class MesureAirMapper {
     /**
      * @param feature    objet réponse de l'api Atmo-France
      * @param dateReleve date du relevé (précisée pour la requete)
@@ -82,7 +62,7 @@ public class MesureMapper {
      * Crée une mesure air
      *
      * @param codeElement code de l'élément récupéré par la requete
-     * @param indice      valeur indexée de la mesure (convention //todo trouvé la convention euro d'indexation
+     * @param indice      valeur indexée de la mesure (convention //todo trouver la convention euro d'indexation
      * @param dateReleve  date du relevé
      * @param dateMaj     date d'insertion
      * @return mesure créée
