+++ src/main/java/fr/diginamic/qualiair/service/MesureService.java
@@ -31,15 +31,4 @@ public class MesureService {
         return mesure;
     }
 
-//    public Mesure findOrCreate(Mesure mesure) {
-//
-//        String key = toKey(mesure);
-//        Mesure existing = cacheService.findInMesureCache(key);
-//        if (existing != null) {
-//            return existing;
-//        }
-//        mesureRepository.save(mesure);
-//        cacheService.putInMesureCache(key, mesure);
-//        return mesure;
-//    }
 }
